+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/SubscriptionsScreen.kt
@@ -1049,7 +1049,6 @@ fun SubscriptionsScreen() {
         }
     }
 
-
     @Composable
     fun SortDialog(onDismissRequest: () -> Unit) {
         var sortingJob = remember<Job?> { null }